fix null entity handling

See merge request binarxsynrgy-2_mainbootcamp/finalprojectsynrgy/team-c/backend!56

+++ src/main/java/com/jalin/jalinappbackend/module/gamification/mission/service/MissionServiceImpl.java
@@ -89,14 +89,11 @@ public class MissionServiceImpl implements MissionService {
                 .orElseThrow(() -> new ResourceNotFoundException("Mission not found"));
         List<UserMission> userMissions = userMissionRepository.findUserMissionsByMission(mission);
 
-        String expiration = mission.getExpiration();
-
-        for (UserMission userMission : userMissions) {
-            userMissionRepository.delete(userMission);
+        if (!userMissions.isEmpty()) {
+            for (UserMission userMission : userMissions) {
+                userMissionRepository.delete(userMission);
+            }
         }
-
         missionRepository.delete(mission);
-
-        userMissionService.forceAssignUserMission(expiration.toUpperCase(Locale.ROOT));
     }
 }
